Rename global to shared
