+++ src/test/shared/gui/QtCoreTest.cpp
