+++ src/test/shared/gui/QtCoreTest.h
