+++ src/test/unittests/CMakeLists.txt
@@ -23,11 +23,11 @@ file(GLOB_RECURSE remove_platform "platform/*")
 list(REMOVE_ITEM headers ${remove_platform})
 list(REMOVE_ITEM sources ${remove_platform})
 
-file(GLOB_RECURSE global_headers "../../test/global/*.h")
-file(GLOB_RECURSE global_sources "../../test/global/*.cpp")
+file(GLOB_RECURSE shared_headers "../../test/shared/*.h")
+file(GLOB_RECURSE shared_sources "../../test/shared/*.cpp")
 
-list(APPEND headers ${global_headers})
-list(APPEND sources ${global_sources})
+list(APPEND headers ${shared_headers})
+list(APPEND sources ${shared_sources})
 
 file(GLOB_RECURSE mock_headers "../../test/mock/*.h")
 file(GLOB_RECURSE mock_sources "../../test/mock/*.cpp")
